placement

+++ GreenFootTetris/Block.java
@@ -1,12 +1,12 @@
 import greenfoot.*;  // (World, Actor, GreenfootImage, Greenfoot and MouseInfo)
 public class Block extends Actor
 {
-    double gridX;
-    double gridY;
+    int gridX;
+    int gridY;
     int worldX;
     int worldY;
 
-    public Block(PieceColor color, double gridX, double gridY) {
+    public Block(PieceColor color, int gridX, int gridY) {
         this(color);
         setGridPosition(gridX, gridY);
     }
@@ -17,7 +17,7 @@ public class Block extends Actor
         setImage(image);
     }
 
-    public void setGridPosition(double x, double y) {
+    public void setGridPosition(int x, int y) {
         gridX = x;
         gridY = y;
         Vector2 gridLocation = MyWorld.posGridToWorld(x, y);
